requirements.txt
pandas>=2.2,<3.0
scikit-learn>=1.5,<2.0
supabase>=2.7,<3.0
python-dotenv>=1.0,<2.0
